Move all c2b functions to the c2b.py file.
