+++ scripts/c2b.py
@@ -1,10 +1,34 @@
 import requests
-from c2b_all import register, simulate
+from scripts.keys import *
+from scripts.access_token import generate_access_token
+from decouple import config
 
-# Run register_url function
+my_access_token = generate_access_token()
 
-# register.register_url() # Only done once
+def register_url():
+    api_url = "https://sandbox.safaricom.co.ke/mpesa/c2b/v1/registerurl"
+    headers = {"Authorization": "Bearer %s" % my_access_token}
+    request = {"ShortCode": c2b_shortcode,
+               "ResponseType": "Completed",
+               "ConfirmationURL": "https://darajavick.herokuapp.com/api/payments/confirmation_url/",
+               "ValidationURL": "https://darajavick.herokuapp.com/api/payments/validation_url/"}
 
-# Run Simulate_c2b_transaction 
+    response = requests.post(api_url, json=request, headers=headers)
 
-simulate.simulate_c2b_transaction()
+    return response.text
+
+
+def simulate_c2b_transaction(phone_number, amount):
+    print('Start excecuting the Simulate C2B function:::')
+    api_url = "https://sandbox.safaricom.co.ke/mpesa/c2b/v1/simulate"
+    headers = {"Authorization": "Bearer %s" % my_access_token}
+
+    request = {"ShortCode": c2b_shortcode,
+               "CommandID": "CustomerPayBillOnline",
+               "Amount": amount,
+               "Msisdn": phone_number,
+               "BillRefNumber": "12345"}
+
+    response = requests.post(api_url, json=request, headers=headers)
+
+    return response.text
